Check-in summary file script
This makes a sample-response by cohort summary file from the raw data in the exacloud sample count directory from 2021-07-24.  
The operative column is "Resampling Step" with 100 scores corresponding to each resampling step size within each cohort.
The purpose is fit to multiple curves for each cohort to allow for the generation of a confidence interval in predicting final accuracy.